Day-8

+++ common-dto-event/src/main/java/com/sk/rk/events/OrchestratorRequestDTO.java
@@ -2,12 +2,10 @@ package com.sk.rk.events;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class OrchestratorRequestDTO {
     private Long customerId;
     private Long productId;
     private Long orderId;
-    private BigDecimal amount;
+    private Double amount;
 }
