+++ customer-service/src/main/java/com/sk/rk/model/entity/Customer.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import nonapi.io.github.classgraph.json.Id;
 
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -18,5 +17,5 @@ public class Customer {
     private String lastName;
     private String email;
     private String password;
-    private BigDecimal balance;
+    private Double balance;
 }
