+++ customer-service/src/main/java/com/sk/rk/repository/CustomerRepositoryImpl.java
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -48,7 +47,7 @@ public class CustomerRepositoryImpl implements CustomerRepository {
                 , rs.getString("last_name")
                 , rs.getString("email")
                 , rs.getString("password")
-                , rs.getBigDecimal("balance")
+                , rs.getDouble("balance")
         ));
     }
 
@@ -61,7 +60,7 @@ public class CustomerRepositoryImpl implements CustomerRepository {
                                 , rs.getString("last_name")
                                 , rs.getString("email")
                                 , rs.getString("password")
-                                , rs.getBigDecimal("balance")
+                                , rs.getDouble("balance")
                         )
                         )
                 );
@@ -72,7 +71,7 @@ public class CustomerRepositoryImpl implements CustomerRepository {
         return jdbcTemplate.queryForObject("select MAX(customer_id) from customer", Long.class);
     }
 
-    public void creditAmount(Long customerId, BigDecimal amount) {
+    public void creditAmount(Long customerId, Double amount) {
 
     }
